+++ src/asset.rs
@@ -297,7 +297,8 @@ pub struct EffectAsset {
     pub simulation_condition: SimulationCondition,
     /// Seed for the pseudo-random number generator.
     ///
-    /// This is uploaded to GPU and used for the various random expressions and quantities computed in shaders.
+    /// This is uploaded to GPU and used for the various random expressions and
+    /// quantities computed in shaders.
     pub prng_seed: u32,
     /// Init modifier defining the effect.
     #[reflect(ignore)]